refactor: add constants for pytests
